refactor : row component - client / server component 분리로 인한 refactor
